withinBounds and checkDomainEquality duplicate functions were removed. These functions now live only in container-helpers.js. Found and replaced all isEqual usages that should be using checkDomainEquality.
